backend/gmail_watch_cron/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
firebase-admin==6.2.0
python-dotenv==1.0.0
pydantic==2.5.0
